requirements.txt
Django==3.0.7
djangorestframework
drf-yasg
channels
requests
Pillow==7.1.2
bs4
celery==4.4.6
redis==3.2.0
channels_redis==2.4.2
django-filter==2.2.0